+++ package.json
@@ -30,7 +30,7 @@
   "dependencies": {
     "@auth0/auth0-react": "^1.8.0",
     "@dataware-tools/api-file-provider-client": "^1.6.2",
-    "@dataware-tools/api-job-store-client": "^1.3.0",
+    "@dataware-tools/api-job-store-client": "^1.4.0",
     "@dataware-tools/api-meta-store-client": "^0.5.1",
     "@dataware-tools/api-permission-manager-client": "^0.2.1",
     "@dataware-tools/app-common": "^21.11.3",
