+++ src/components/molecules/FilePreviewer/RosbagPreviewer/index.tsx
@@ -1,38 +1,15 @@
-import { useAuth0 } from "@auth0/auth0-react";
-import { jobStore } from "@dataware-tools/api-job-store-client";
-import {
-  ErrorMessageProps,
-  DialogTitle,
-  ErrorMessage,
-  extractErrorMessageFromFetchError,
-} from "@dataware-tools/app-common";
+import { DialogTitle } from "@dataware-tools/app-common";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Tab from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { JobSubmitter } from "./JobSubmitter";
-import { JobViewer } from "./JobViewer";
-import {
-  fetchJobStore,
-  useGetJobTemplate,
-  useGetJobTypes,
-  useListJobTemplate,
-} from "utils";
 
 type Props = {
   tabIndex: number;
   handleChangeTab: (_: React.SyntheticEvent, newValue: number) => void;
-  url: string;
-  error?: ErrorMessageProps;
-  onChangeJobTemplate: (jobTemplateId: string) => void;
-  onSubmitJob: () => void;
-  jobTemplateId?: string;
-  jobTemplateList?: jobStore.JobTemplatesModel["job_templates"];
-  jobTemplate?: jobStore.JobTemplateModel;
-  jobType?: jobStore.JobTypeModel;
-  job?: jobStore.JobPostedModel;
-};
+} & RosbagPreviewerProps;
 
 export type RosbagPreviewerProps = {
   databaseId: string;
@@ -44,18 +21,11 @@ const RosbagPreviewerPresentation = ({
   tabIndex,
   handleChangeTab,
   url,
-  error,
-  onChangeJobTemplate,
-  onSubmitJob,
-  jobTemplateId,
-  jobTemplateList,
-  jobTemplate,
-  jobType,
-  job,
+  databaseId,
+  recordId,
+  filePath,
 }: Props): JSX.Element => {
-  return error ? (
-    <ErrorMessage {...error} />
-  ) : (
+  return (
     <Box>
       <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
         <Tabs value={tabIndex} onChange={handleChangeTab} aria-label="tabs">
@@ -80,18 +50,11 @@ const RosbagPreviewerPresentation = ({
       {tabIndex === 1 && (
         <Box>
           <DialogTitle>Submit a job</DialogTitle>
-          {jobTemplateList && (
-            <JobSubmitter
-              jobTemplateId={jobTemplateId}
-              jobTemplateList={jobTemplateList}
-              jobTemplate={jobTemplate}
-              jobType={jobType}
-              job={job}
-              onSubmitJob={onSubmitJob}
-              onChangeJobTemplate={onChangeJobTemplate}
-            />
-          )}
-          {jobType && job && <JobViewer jobType={jobType} job={job} />}
+          <JobSubmitter
+            databaseId={databaseId}
+            recordId={recordId}
+            filePath={filePath}
+          />
         </Box>
       )}
     </Box>
@@ -104,63 +67,7 @@ export const RosbagPreviewer = ({
   filePath,
   url,
 }: RosbagPreviewerProps): JSX.Element => {
-  const { getAccessTokenSilently: getAccessToken } = useAuth0();
   const [tabIndex, setTabIndex] = useState<number>(0);
-  const [error, setError] = useState<ErrorMessageProps | undefined>(undefined);
-  const [jobTemplateId, setJobTemplateId] = useState<string | undefined>(
-    undefined
-  );
-  const [job, setJob] = useState<jobStore.JobPostedModel | undefined>(
-    undefined
-  );
-  const { data: listJobTemplateRes, error: listJobTemplateError } =
-    useListJobTemplate(getAccessToken, undefined as never);
-  const { data: getJobTemplateRes, error: getJobTemplateError } =
-    useGetJobTemplate(getAccessToken, {
-      jobTemplateId: jobTemplateId ? parseInt(jobTemplateId, 10) : undefined,
-    });
-  const { data: getJobTypeRes, error: getJobTypeError } = useGetJobTypes(
-    getAccessToken,
-    {
-      jobTypeUid: getJobTemplateRes?.job_template.job_type_uid,
-    }
-  );
-
-  const fetchError =
-    listJobTemplateError || getJobTemplateError || getJobTypeError;
-  useEffect(() => {
-    if (fetchError) {
-      const { reason, instruction } =
-        extractErrorMessageFromFetchError(fetchError);
-      setError({ reason, instruction });
-    } else {
-      setError(undefined);
-    }
-  }, [fetchError]);
-
-  const onSubmitJob = async () => {
-    if (!getJobTemplateRes) {
-      return;
-    }
-    const [data, error] = await fetchJobStore(
-      getAccessToken,
-      jobStore.JobService.createJob,
-      {
-        requestBody: {
-          job_template_id: parseInt(getJobTemplateRes.job_template_id),
-          database_id: databaseId,
-          record_id: recordId,
-          path_to_rosbag: filePath, // FIXME: Generalize JobPostModel
-        },
-      }
-    );
-    if (error) {
-      const { reason, instruction } = extractErrorMessageFromFetchError(error);
-      setError({ reason, instruction });
-    } else {
-      setJob(data);
-    }
-  };
 
   const handleChangeTab = (_: React.SyntheticEvent, newValue: number) => {
     setTabIndex(newValue);
@@ -171,14 +78,9 @@ export const RosbagPreviewer = ({
       tabIndex={tabIndex}
       handleChangeTab={handleChangeTab}
       url={url}
-      error={error}
-      jobTemplateId={jobTemplateId}
-      onChangeJobTemplate={setJobTemplateId}
-      onSubmitJob={onSubmitJob}
-      jobTemplate={getJobTemplateRes}
-      jobTemplateList={listJobTemplateRes?.job_templates}
-      jobType={getJobTypeRes}
-      job={job}
+      databaseId={databaseId}
+      recordId={recordId}
+      filePath={filePath}
     />
   );
 };
